timer double

+++ src/main/java/org/overrun/real2d/client/Timer.java
@@ -11,7 +11,7 @@ public class Timer {
     public static final long MAX_NS_PER_UPDATE = 1000000000L;
     public static final int MAX_TICKS_PER_UPDATE = 100;
     private final double tps;
-    private long lastTime;
+    private double lastTime;
     public int ticks = 0;
     public double delta = 0;
     public double fps = 0;
@@ -19,19 +19,19 @@ public class Timer {
 
     public Timer(double tps) {
         this.tps = tps;
-        lastTime = (long) (glfwGetTime() * NS_PER_SECOND);
+        lastTime = glfwGetTime() * NS_PER_SECOND;
     }
 
     public void advanceTime() {
-        long now = (long) (glfwGetTime() * NS_PER_SECOND);
-        long passedNs = now - lastTime;
+        double now = glfwGetTime() * NS_PER_SECOND;
+        double passedNs = now - lastTime;
         lastTime = now;
         if (passedNs < 0) {
             passedNs = 0;
         } else if (passedNs > MAX_NS_PER_UPDATE) {
             passedNs = MAX_NS_PER_UPDATE;
         }
-        fps = (double) (NS_PER_SECOND / passedNs);
+        fps = NS_PER_SECOND / passedNs;
         passedTime += passedNs * tps / NS_PER_SECOND;
         ticks = (int) passedTime;
         if (ticks > MAX_TICKS_PER_UPDATE) {
